edit destroy.sh

+++ destroy.sh
@@ -13,6 +13,19 @@ fi
 
 echo "NOTE: Games image is $games_image"
 
+desktop_image=$(gcloud compute images list \
+  --filter="name~'^desktop-image' AND family=desktop-images" \
+  --sort-by="~creationTimestamp" \
+  --limit=1 \
+  --format="value(name)")
+
+if [[ -z "$desktop_image" ]]; then
+  echo "ERROR: No latest image found for 'desktop-image' in family 'desktop-images'."
+  exit 1
+fi
+
+echo "NOTE: Desktop image is $desktop_image"
+
 cd 03-deploy
 terraform init
 terraform destroy -var="games_image_name=$games_image" -auto-approve
@@ -34,6 +47,7 @@ else
     gcloud compute images delete "$image" --quiet || echo "WARNING: Failed to delete image: $image"
   done
 fi
+
 
 cd 01-infrastructure
 terraform init
